% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopMI-package.R
\docType{package}
\name{CopMI-package}
\alias{CopMI}
\alias{CopMI-package}
\title{CopMI: Gaussian copula imputation for left-censored exposures}
\description{
Use \code{\link[=copula_em_impute]{copula_em_impute()}} for multiple imputation,
\code{\link[=complete]{complete()}} to extract completed data, and
\code{\link[=diagnostics]{diagnostics()}} to inspect model diagnostics.
}
\seealso{
\link{nhanes_pah}, \code{vignette("copmi-workflow", package = "CopMI")}
}
\author{
\strong{Maintainer}: Xuan Han \email{1571836967@qq.com}

Authors:
\itemize{
  \item Xuan Han \email{1571836967@qq.com}
}

}
\keyword{internal}
